decimal -> double fixes + small .gitignore update :)

+++ AI_NETCORE_API/Domain/Repositories/BuyOfferRepo/Abstract/IBuyOfferRepository.cs
@@ -11,7 +11,7 @@ namespace Domain.Repositories.BuyOfferRepo.Abstract
     public interface IBuyOfferRepository : IRepositoryBase<BuyOffer>
     {
         RepositoryResponse<IEnumerable<BusinessObject.BuyOffer>> GetBuyOffersByUserId(int id);
-        long CreateBuyOffer(int companyId, int amount, decimal price);
+        long CreateBuyOffer(int companyId, int amount, double price);
         long WithdrawBuyOffer(int buyOfferId);
     }
 }
